+++ frontend/components/mindmap-component/MindMapCategory.js
