+++ frontend/components/notes-component/CurrentFolderContainer.js
