+++ frontend/components/notes-component/Folder.js
