+++ frontend/components/notes-component/NoteEditor.js
