+++ frontend/components/notes-component/NotesScreen.js
